balanceviewModel ended,balanceFragment ended,balanceadapter ended without firstInssert and bug

+++ app/src/main/java/com/imranmelikov/zipex/model/AdminLink.kt
@@ -10,7 +10,7 @@ data class AdminLink(val url:String,
                      ,val count:Int
                      ,val color:String
                      ,val size:String
-                     ,val price:Int
+                     ,val price:Double
                      ,val comment:String
                      ,val history:String
                      ,val marketName:String
