+++ app/src/main/java/com/imranmelikov/zipex/model/BalanceTotalAzn.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class BalanceTotalAzn(val balanceTotal:Int) {
+data class BalanceTotalAzn(val balanceTotal:Double) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
