+++ app/src/main/java/com/imranmelikov/zipex/model/BalanceTry.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 import java.util.Date
 
 @Entity
-data class BalanceTry(val history:String,val amount:Int,val balance:Int) {
+data class BalanceTry(val history:String,val amount:Double,val balance:Double) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
