+++ app/src/main/java/com/imranmelikov/zipex/model/DebtTotal.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class DebtTotal(val debtTotal:Int) {
+data class DebtTotal(val debtTotal:Double) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
