+++ app/src/main/java/com/imranmelikov/zipex/model/Link.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 import java.util.Date
 
 @Entity
-data class Link(val url:String,val category:String,val count:Int,val color:String,val size:String,val price:Int,val comment:String,val history:String) {
+data class Link(val url:String,val category:String,val count:Int,val color:String,val size:String,val price:Double,val comment:String,val history:String) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
