+++ app/src/main/java/com/imranmelikov/zipex/model/Order1.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 import java.util.Date
 
 @Entity
-data class Order1(val url:String,val category:String,val count:Int,val color:String,val size:String,val price:Int,val comment:String,val history:String,val marketName:String,val marketCode:String,val office:String) {
+data class Order1(val url:String,val category:String,val count:Int,val color:String,val size:String,val price:Double,val comment:String,val history:String,val marketName:String,val marketCode:String,val office:String) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
