+++ app/src/main/java/com/imranmelikov/zipex/model/Order5.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 import java.util.Date
 
 @Entity
-data class Order5(val url:String, val category:String, val count:Int, val color:String, val size:String, val price:Int, val comment:String, val history:String, val marketName:String, val marketCode:String, val office:String,val delivery:Int,val weight:Int) {
+data class Order5(val url:String, val category:String, val count:Int, val color:String, val size:String, val price:Double, val comment:String, val history:String, val marketName:String, val marketCode:String, val office:String,val delivery:Int,val weight:Int) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
