+++ app/src/main/java/com/imranmelikov/zipex/model/ZipexMoneyDepot.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class ZipexMoneyDepot(val moneyDepot:Int) {
+data class ZipexMoneyDepot(val moneyDepot:Double) {
     @PrimaryKey(autoGenerate = true)
     var uuid:Int?=null
 }
